Added Tasks Page

+++ src/pages/api/projects/[id]/team/[teamId].ts
@@ -35,8 +35,31 @@ export default auth()
 		});
 	})
 	.delete(async (req, res) => {
+		// Find the team member
+		const member = await prisma.projectTeam.findUnique({
+			where: { id: req.query.teamId as string },
+		});
+
+		if (!member) {
+			return res.status(404).json({
+				status: 'error',
+				message: 'Team member with specified ID was not found!',
+			});
+		}
+
+		// Delete/Remove the team member
 		await prisma.projectTeam.delete({
 			where: { id: req.query.teamId as string },
+		});
+
+		// Delete/Remove all task following associated with the team member
+		await prisma.projectTaskFollower.deleteMany({
+			where: {
+				employeeId: member.employeeId,
+				task: {
+					projectId: req.query.id as string,
+				},
+			},
 		});
 		return res.status(200).json({
 			status: 'success',
